+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 ### Added
 - The `SubmitButton` component, used within `Form` components.
+- The ability to set a `validator` prop on form fields that should either return an error message string or `null`.
+
+### Changed
+- All of the form fields no longer track the value in state, they rely on being controlled components.
+- The value of the label for form fields is now set using `children`.
+- `Form` components now fully take over the `onChange` value of various form fields and use it to track state.
 
 ## [2.3.0] - 2018-08-21
 ### Added